+++ src/main/java/net/silentchaos512/gems/compat/gear/SGearStatHandler.java
@@ -24,6 +24,7 @@ import net.minecraft.item.ItemStack;
 import net.minecraft.util.text.TextFormatting;
 import net.minecraftforge.event.RegistryEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
+import net.silentchaos512.gear.api.event.GetMaterialStatsEvent;
 import net.silentchaos512.gear.api.event.GetStatModifierEvent;
 import net.silentchaos512.gear.api.stats.ItemStat;
 import net.silentchaos512.gear.api.stats.ItemStats;
@@ -31,7 +32,10 @@ import net.silentchaos512.gear.api.stats.StatInstance;
 import net.silentchaos512.gems.SilentGems;
 import net.silentchaos512.gems.init.GemsEnchantments;
 
+import java.util.Collections;
+import java.util.EnumSet;
 import java.util.Map;
+import java.util.Set;
 import java.util.function.ToDoubleFunction;
 
 /**
@@ -49,7 +53,7 @@ public class SGearStatHandler {
                     .put(ItemStats.ENCHANTABILITY,    prop -> prop.originalStat * (1 + prop.superchargedLevel * (Math.sqrt(prop.chargeability) - 1)))
                     .put(ItemStats.RARITY,            prop -> prop.originalStat + prop.superchargedLevel * 10)
                     .put(ItemStats.HARVEST_LEVEL,     prop -> prop.originalStat + prop.superchargedLevel)
-                    .put(ItemStats.HARVEST_SPEED,     prop -> prop.originalStat + prop.superchargedLevel * prop.chargeValue * 3)
+                    .put(ItemStats.HARVEST_SPEED,     prop -> prop.originalStat + prop.superchargedLevel * prop.chargeValue * 2)
                     .put(ItemStats.MELEE_DAMAGE,      prop -> prop.originalStat + prop.chargeValue * 2)
                     .put(ItemStats.MAGIC_DAMAGE,      prop -> prop.originalStat + prop.chargeValue * 2)
                     .put(ItemStats.RANGED_DAMAGE,     prop -> prop.originalStat + prop.chargeValue)
@@ -59,12 +63,47 @@ public class SGearStatHandler {
                     //@formatter:on
                     .build();
 
+    private static final Set<StatInstance.Operation> SUPPORTED_OPS = Collections.unmodifiableSet(EnumSet.of(
+            StatInstance.Operation.AVG,
+            StatInstance.Operation.MAX,
+            StatInstance.Operation.ADD
+    ));
+
     public static void registerStats(RegistryEvent.Register<ItemStat> event) {
         event.getRegistry().register(CHARGEABILITY.setRegistryName(SilentGems.getId("chargeability")));
     }
 
     @SubscribeEvent
+    public void onGetMaterialStats(GetMaterialStatsEvent event) {
+        ItemStack stack = event.getMaterial().getItem();
+        int supercharged = EnchantmentHelper.getEnchantmentLevel(GemsEnchantments.SUPERCHARGED.get(), stack);
+
+        if (supercharged > 0 && BOOSTED_STATS.containsKey(event.getStat())) {
+            float chargeability = CHARGEABILITY.compute(0, event.getMaterial().getStatModifiers(CHARGEABILITY, event.getPartType(), ItemStack.EMPTY));
+            if (chargeability == 0) chargeability = 1;
+            float chargeLevel = chargeability * supercharged;
+
+            for (int i = 0; i < event.getModifiers().size(); ++i) {
+                StatInstance instance = event.getModifiers().get(i);
+                if (isSupportedOp(instance)) {
+                    // Replace instance with a modified one
+                    ChargedProperties chargedProperties = new ChargedProperties(supercharged, chargeLevel, instance.getValue());
+                    float statValue = (float) BOOSTED_STATS.get(event.getStat()).applyAsDouble(chargedProperties);
+                    StatInstance replacement = new StatInstance(statValue, instance.getOp());
+                    event.getModifiers().remove(instance);
+                    event.getModifiers().add(i, replacement);
+                }
+            }
+        }
+    }
+
+    private static boolean isSupportedOp(StatInstance instance) {
+        return SUPPORTED_OPS.contains(instance.getOp());
+    }
+
+    @SubscribeEvent
     public void onGetPartStats(GetStatModifierEvent event) {
+        // TODO: Remove me! Left in for legacy simple mains
         ItemStack stack = event.getPart().getCraftingItem();
         int supercharged = EnchantmentHelper.getEnchantmentLevel(GemsEnchantments.SUPERCHARGED.get(), stack);